Finished implementing two more user stories
First:
As an organizer adding or removin player to a yet to be chosen team, I
should be prompted with an alphabetically ordered list of teams to
choose from, so that I can quickly locate the team and avoid typos.

Second:
As an organizer adding or removing a player to a chosen team, I should
be prompted with an alphabetically ordered list of players along with
their stats, so that i can quickly locate player and take action.

Implemented by using Collections.sort() and by implementing Comparable
interface by Team class.