Implemented League Balance Report
As an organizer who is planning teams, I should be able to see a League
Balance Report for all teams in the league showing total count of
experienced vs inexperienced players, so i can determine from a high
level if the teams are fairly balanced regarding prevoius experience.

To implement used map of maps - commented for clarity in code.